routers: move func used for web and api to common
