+++ routers/api/v1/repo/file.go
@@ -17,7 +17,7 @@ import (
 	"code.gitea.io/gitea/modules/repofiles"
 	api "code.gitea.io/gitea/modules/structs"
 	"code.gitea.io/gitea/modules/web"
-	"code.gitea.io/gitea/routers/web/repo"
+	"code.gitea.io/gitea/routers/common"
 )
 
 // GetRawFile get a file by path on a repository
@@ -83,7 +83,7 @@ func GetRawFile(ctx *context.APIContext) {
 		}
 		return
 	}
-	if err = repo.ServeBlob(ctx.Context, blob); err != nil {
+	if err = common.ServeBlob(ctx.Context, blob); err != nil {
 		ctx.Error(http.StatusInternalServerError, "ServeBlob", err)
 	}
 }
@@ -126,7 +126,7 @@ func GetArchive(ctx *context.APIContext) {
 	ctx.Repo.GitRepo = gitRepo
 	defer gitRepo.Close()
 
-	repo.Download(ctx.Context)
+	common.Download(ctx.Context)
 }
 
 // GetEditorconfig get editor config of a repository
